+++ optimacore/project.py
@@ -8,10 +8,10 @@ The archetypal example is an epidemic within a geographical location, where enti
 from optimacore.system import SystemSettings as SS
 from optimacore.excel import ExcelSettings as ES
 
-from optimacore.system import applyToAllMethods, logUsage, accepts, returns, logger
+from optimacore.system import applyToAllMethods, logUsage, accepts
 from optimacore.framework import ProjectFramework
 from optimacore.data import ProjectData
-from optimacore.workbook import writeWorkbook
+from optimacore.workbook_export import writeWorkbook
 
 @applyToAllMethods(logUsage)
 class Project(object):
@@ -41,22 +41,3 @@ class Project(object):
     def setName(self, name):
         """ Set primary human-readable identifier for the project. """
         self.name = name
-    
-    #@returns(str)
-    #def getName(self):
-    #    """ Get primary human-readable identifier for the project. """
-    #    return self.name
-    
-    #@accepts(ProjectFramework)
-    #def setFramework(self, framework):
-    #    """ Set the underlying context framework for the project. """
-    #    self.framework = framework
-    
-    #@returns(ProjectFramework)
-    #def getFramework(self):
-    #    """ Get the underlying context framework for the project. """
-    #    return self.framework
-
-#    def __repr__(self):
-#        """ String representation of the project. """
-#        return ""